+++ src/components/Header.css
@@ -1,0 +1,15 @@
+nav {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    max-width: 1200px;
+    height: 70px;
+    margin: auto;
+    padding: 0px 10px;
+    font-size: 20px;
+    border: 1px solid #d7d7d7d7;
+}
+
+.logo {
+    max-width: 45px;
+}
