+++ src/components/StudentList.css
@@ -1,0 +1,34 @@
+ul {
+    max-width: 800px;
+    margin: 20px auto;
+    padding: 20px;
+    border: 1px solid #d7d7d7;
+    border-radius: 5px;
+}
+
+h1 {
+    font-size: 28px;
+    text-align: center;
+}
+
+.header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 20px 10px;
+}
+
+.delete {
+    background: red;
+}
+
+li {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    box-shadow: rgba(0, 0, 0, 0.16) 0px 1px 4px;
+    margin: 20px 5px;
+    padding: 25px;
+    font-size: 20px;
+    border-radius: 5px;
+}
